chore: rename `aranet4` integration to `aranet`
