+++ homeassistant/components/aranet/const.py
@@ -1,0 +1,3 @@
+"""Constants for the Aranet integration."""
+
+DOMAIN = "aranet"
