+++ homeassistant/components/aranet/sensor.py
