+++ homeassistant/components/aranet/strings.json
