+++ homeassistant/components/aranet/translations/en.json
