--- homeassistant/components/aranet4/const.py
@@ -1,3 +1,0 @@
-"""Constants for the Aranet4 integration."""
-
-DOMAIN = "aranet4"
