+++ requirements_all.txt
@@ -335,7 +335,7 @@ aprslib==0.7.0
 # homeassistant.components.aqualogic
 aqualogic==2.6
 
-# homeassistant.components.aranet4
+# homeassistant.components.aranet
 aranet4==2.1.3
 
 # homeassistant.components.arcam_fmj
