+++ requirements_test_all.txt
@@ -298,7 +298,7 @@ apprise==1.1.0
 # homeassistant.components.aprs
 aprslib==0.7.0
 
-# homeassistant.components.aranet4
+# homeassistant.components.aranet
 aranet4==2.1.3
 
 # homeassistant.components.arcam_fmj
