+++ tests/components/aranet/test_config_flow.py
@@ -1,8 +1,8 @@
-"""Test the Aranet4 config flow."""
+"""Test the Aranet config flow."""
 from unittest.mock import patch
 
 from homeassistant import config_entries
-from homeassistant.components.aranet4.const import DOMAIN
+from homeassistant.components.aranet.const import DOMAIN
 from homeassistant.core import HomeAssistant
 from homeassistant.data_entry_flow import FlowResultType
 
@@ -25,7 +25,7 @@ async def test_async_step_bluetooth_valid_device(hass):
     )
     assert result["type"] == FlowResultType.FORM
     assert result["step_id"] == "bluetooth_confirm"
-    with patch("homeassistant.components.aranet4.async_setup_entry", return_value=True):
+    with patch("homeassistant.components.aranet.async_setup_entry", return_value=True):
         result2 = await hass.config_entries.flow.async_configure(
             result["flow_id"], user_input={}
         )
@@ -92,7 +92,7 @@ async def test_async_step_user_takes_precedence_over_discovery(hass):
     assert result["step_id"] == "bluetooth_confirm"
 
     with patch(
-        "homeassistant.components.aranet4.config_flow.async_discovered_service_info",
+        "homeassistant.components.aranet.config_flow.async_discovered_service_info",
         return_value=[VALID_DATA_SERVICE_INFO],
     ):
         result = await hass.config_entries.flow.async_init(
@@ -101,7 +101,7 @@ async def test_async_step_user_takes_precedence_over_discovery(hass):
         )
         assert result["type"] == FlowResultType.FORM
 
-    with patch("homeassistant.components.aranet4.async_setup_entry", return_value=True):
+    with patch("homeassistant.components.aranet.async_setup_entry", return_value=True):
         result2 = await hass.config_entries.flow.async_configure(
             result["flow_id"],
             user_input={"address": "aa:bb:cc:dd:ee:ff"},
@@ -128,7 +128,7 @@ async def test_async_step_user_no_devices_found(hass: HomeAssistant):
 async def test_async_step_user_only_other_devices_found(hass: HomeAssistant):
     """Test setup from service info cache with only other devices found."""
     with patch(
-        "homeassistant.components.aranet4.config_flow.async_discovered_service_info",
+        "homeassistant.components.aranet.config_flow.async_discovered_service_info",
         return_value=[NOT_ARANET4_SERVICE_INFO],
     ):
         result = await hass.config_entries.flow.async_init(
@@ -142,7 +142,7 @@ async def test_async_step_user_only_other_devices_found(hass: HomeAssistant):
 async def test_async_step_user_with_found_devices(hass: HomeAssistant):
     """Test setup from service info cache with devices found."""
     with patch(
-        "homeassistant.components.aranet4.config_flow.async_discovered_service_info",
+        "homeassistant.components.aranet.config_flow.async_discovered_service_info",
         return_value=[VALID_DATA_SERVICE_INFO],
     ):
         result = await hass.config_entries.flow.async_init(
@@ -151,7 +151,7 @@ async def test_async_step_user_with_found_devices(hass: HomeAssistant):
         )
     assert result["type"] == FlowResultType.FORM
     assert result["step_id"] == "user"
-    with patch("homeassistant.components.aranet4.async_setup_entry", return_value=True):
+    with patch("homeassistant.components.aranet.async_setup_entry", return_value=True):
         result2 = await hass.config_entries.flow.async_configure(
             result["flow_id"],
             user_input={"address": "aa:bb:cc:dd:ee:ff"},
@@ -165,7 +165,7 @@ async def test_async_step_user_with_found_devices(hass: HomeAssistant):
 async def test_async_step_user_device_added_between_steps(hass: HomeAssistant):
     """Test the device gets added via another flow between steps."""
     with patch(
-        "homeassistant.components.aranet4.config_flow.async_discovered_service_info",
+        "homeassistant.components.aranet.config_flow.async_discovered_service_info",
         return_value=[VALID_DATA_SERVICE_INFO],
     ):
         result = await hass.config_entries.flow.async_init(
@@ -181,7 +181,7 @@ async def test_async_step_user_device_added_between_steps(hass: HomeAssistant):
     )
     entry.add_to_hass(hass)
 
-    with patch("homeassistant.components.aranet4.async_setup_entry", return_value=True):
+    with patch("homeassistant.components.aranet.async_setup_entry", return_value=True):
         result2 = await hass.config_entries.flow.async_configure(
             result["flow_id"],
             user_input={"address": "aa:bb:cc:dd:ee:ff"},
@@ -199,7 +199,7 @@ async def test_async_step_user_with_found_devices_already_setup(hass: HomeAssist
     entry.add_to_hass(hass)
 
     with patch(
-        "homeassistant.components.aranet4.config_flow.async_discovered_service_info",
+        "homeassistant.components.aranet.config_flow.async_discovered_service_info",
         return_value=[VALID_DATA_SERVICE_INFO],
     ):
         result = await hass.config_entries.flow.async_init(
@@ -213,7 +213,7 @@ async def test_async_step_user_with_found_devices_already_setup(hass: HomeAssist
 async def test_async_step_user_old_firmware(hass: HomeAssistant):
     """Test we can't set up a device with firmware too old to report measurements."""
     with patch(
-        "homeassistant.components.aranet4.config_flow.async_discovered_service_info",
+        "homeassistant.components.aranet.config_flow.async_discovered_service_info",
         return_value=[OLD_FIRMWARE_SERVICE_INFO],
     ):
         result = await hass.config_entries.flow.async_init(
@@ -222,7 +222,7 @@ async def test_async_step_user_old_firmware(hass: HomeAssistant):
         )
     assert result["type"] == FlowResultType.FORM
     assert result["step_id"] == "user"
-    with patch("homeassistant.components.aranet4.async_setup_entry", return_value=True):
+    with patch("homeassistant.components.aranet.async_setup_entry", return_value=True):
         result2 = await hass.config_entries.flow.async_configure(
             result["flow_id"],
             user_input={"address": "aa:bb:cc:dd:ee:ff"},
@@ -234,7 +234,7 @@ async def test_async_step_user_old_firmware(hass: HomeAssistant):
 async def test_async_step_user_integrations_disabled(hass: HomeAssistant):
     """Test we can't set up a device the device's integration setting disabled."""
     with patch(
-        "homeassistant.components.aranet4.config_flow.async_discovered_service_info",
+        "homeassistant.components.aranet.config_flow.async_discovered_service_info",
         return_value=[DISABLED_INTEGRATIONS_SERVICE_INFO],
     ):
         result = await hass.config_entries.flow.async_init(
@@ -243,7 +243,7 @@ async def test_async_step_user_integrations_disabled(hass: HomeAssistant):
         )
     assert result["type"] == FlowResultType.FORM
     assert result["step_id"] == "user"
-    with patch("homeassistant.components.aranet4.async_setup_entry", return_value=True):
+    with patch("homeassistant.components.aranet.async_setup_entry", return_value=True):
         result2 = await hass.config_entries.flow.async_configure(
             result["flow_id"],
             user_input={"address": "aa:bb:cc:dd:ee:ff"},
